+++ src/rows.cpp
@@ -1,6 +1,7 @@
 #include <Rcpp.h>
 #include <dplyr.h>
 #include "map.h"
+#include "utils.h"
 #include "rows-data.h"
 #include "rows-formatter.h"
 
@@ -69,6 +70,31 @@ extern "C" SEXP invoke_rows_impl(SEXP env, SEXP d_name_, SEXP f_name_) {
   BEGIN_RCPP
   // Map in parallel over the rows of the data frame
   SEXP results = PROTECT(pmap_impl(env, d_name_, f_name_, Rf_mkChar("list")));
+
+  // Create the output data frame
+  results = PROTECT(Slices::process_slices(results, env));
+
+  UNPROTECT(2);
+  return results;
+  END_RCPP
+}
+
+extern "C" SEXP map_by_slice_impl(SEXP env, SEXP d_name_, SEXP f_name_) {
+  BEGIN_RCPP
+  const char* d_name = CHAR(Rf_asChar(d_name_));
+  SEXP d = Rf_install(d_name);
+  SEXP d_val = Rf_eval(d, env);
+
+  // Create list of data frames, one by slice
+  SEXP results = PROTECT(Slices::subset_slices(d_val));
+
+  // Map over those lists
+  for (int i = 0; i < Rf_length(results); ++i) {
+    Rf_defineVar(d, get_vector_elt(results, i), env);
+    SEXP result = PROTECT(map_impl(env, d_name_, f_name_, Rf_mkChar("list")));
+    set_vector_elt(results, i, as_data_frame(result));
+    UNPROTECT(1);
+  }
 
   // Create the output data frame
   results = PROTECT(Slices::process_slices(results, env));